# Add support for registering handlers for different 404 routes